An Excel tool for converting EDL files
EDL files can be created from a timeline when editing video files. This spreadsheet tool will import such an EDL file and transform it into a text format that can be used directly in YouTube as a way of producing Chapter markings. In addition the spreadsheet produces a markup code snippet (with the same chapter content) that can be pasted into a Wiki for example the MERG knowledge base.  The final step (copying the text produced by this tool and pasting in the destination) is up to the user.